+++ src/main/java/org/cube/converter/util/legacy/RotationUtil.java
@@ -11,10 +11,10 @@ import java.util.HashMap;
 // Credit goes to BlockBench. https://github.com/JannisX11/blockbench
 public class RotationUtil {
     public static void rotateIfPossible(final Cube cube) {
-        double[] array = cube.getRotation().toArray();
+        float[] array = cube.getRotation().toArray();
 
         for (int axis = 0; axis < array.length; axis++) {
-            double rotation = array[axis];
+            float rotation = array[axis];
             if (axis != 2) {
                 rotation = -rotation;
             }
@@ -35,9 +35,9 @@ public class RotationUtil {
 
             RotationUtil.rotate90Degrees(cube, rotation, axis, true);
 
-            double[] rotAxes = ArrayUtil.clone(array);
+            float[] rotAxes = ArrayUtil.clone(array);
             var i = 0;
-            Double temp_rot = null;
+            Float temp_rot = null;
             Integer temp_i = null;
             while (i < 3) {
                 if (i != axis) {
@@ -70,17 +70,17 @@ public class RotationUtil {
         array = cube.getRotation().toArray();
 
         int axis = cube.getAxis().equals("x") ? 0 : cube.getAxis().equals("y") ? 1 : 2;
-        double actual = array[axis];
+        float actual = array[axis];
         if (axis != 2) actual = -actual;
 
-        final double[] pivotArray = cube.getPivot().toArray();
+        final float[] pivotArray = cube.getPivot().toArray();
         if (Math.abs(actual) == 135 && pivotArray[0] == 0 && pivotArray[1] == 0 && pivotArray[2] == 0) {
-            final double newAngle = actual > 0 ? -45 : 45;
+            final float newAngle = actual > 0 ? -45 : 45;
             cube.getRotation().set(axis == 0 ? newAngle : 0, axis == 1 ? newAngle : 0, axis == 2 ? newAngle : 0);
             return;
         }
 
-        double hackyRotation = 90 - Math.abs(actual);
+        float hackyRotation = 90 - Math.abs(actual);
         boolean isHackyValid = MathUtil.isValidJavaAngle(hackyRotation) && !MathUtil.isValidJavaAngle(actual)
                 && actual != 0D && hackyRotation != 0D;
 
@@ -88,17 +88,17 @@ public class RotationUtil {
             return;
         }
 
-        double mulValue = Math.abs(actual) / actual;
+        float mulValue = Math.abs(actual) / actual;
         hackyRotation = (Math.abs(hackyRotation) * mulValue);
 
         RotationUtil.rotate90Degrees(cube, 90 * mulValue, axis, true);
 
-        final double newAngle = Math.abs(actual) > 90 || Math.abs(actual) == 112.5 ? hackyRotation : - hackyRotation;
+        final float newAngle = Math.abs(actual) > 90 || Math.abs(actual) == 112.5 ? hackyRotation : - hackyRotation;
         cube.getRotation().set(axis == 0 ? newAngle : 0, axis == 1 ? newAngle : 0, axis == 2 ? newAngle : 0);
     }
 
-    public static void rotate90Degrees(final Cube element, double rotation, int axis, boolean updateUv) {
-        final double[] pivotArray = element.getPivot().toArray();
+    public static void rotate90Degrees(final Cube element, float rotation, int axis, boolean updateUv) {
+        final float[] pivotArray = element.getPivot().toArray();
 
         switch (axis) {
             case 0 -> {
@@ -121,18 +121,18 @@ public class RotationUtil {
         }
     }
 
-    private static void rotate90Degrees(final Cube element, int axis, int steps, double[] origin, boolean updateUv) {
+    private static void rotate90Degrees(final Cube element, int axis, int steps, float[] origin, boolean updateUv) {
         origin = ArrayUtil.clone(origin);
         origin = ArrayUtil.javaOffsetArray(origin);
         // origin[0] = -origin[0];
 
-        double[] from = element.getPosition().asJavaPosition(element.getSize()).toArray();
-        double[] to = ArrayUtil.add(from, element.getSize().toArray());
+        float[] from = element.getPosition().asJavaPosition(element.getSize()).toArray();
+        float[] to = ArrayUtil.add(from, element.getSize().toArray());
 
         while (steps > 0) {
             steps--;
             //Swap coordinate thingy
-            double[] cloneTo = ArrayUtil.clone(to);
+            float[] cloneTo = ArrayUtil.clone(to);
              switch(axis) {
                 case 0, 1 -> {
                     to[2] = from[2];
@@ -148,7 +148,7 @@ public class RotationUtil {
             rotateCoord(to, axis, origin);
             // element.origin(rotateCoord(element.origin(), axis, origin));
 
-            double[] size = ArrayUtil.size(to, from);
+            float[] size = ArrayUtil.size(to, from);
             element.getSize().set(new Position3V(size));
             element.getPosition().set(new Position3V(from).asBedrockPosition(element.getSize()));
 
@@ -156,7 +156,7 @@ public class RotationUtil {
                 continue;
             }
 
-            final HashMap<Direction, Double[]> old = new HashMap<>(element.getUvMap().getMap());
+            final HashMap<Direction, Float[]> old = new HashMap<>(element.getUvMap().getMap());
 
             switch (axis) {
                 case 0 -> {
@@ -183,12 +183,12 @@ public class RotationUtil {
         }
     }
 
-    private static void rotateCoord(double[] array, int axis, double[] origin) {
-        Double a = null;
+    private static void rotateCoord(float[] array, int axis, float[] origin) {
+        Float a = null;
         int b = -1;
 
         for (int i = 0; i < array.length; i++) {
-            double s = array[i];
+            float s = array[i];
             if (i != axis) {
                 if (a == null) {
                     a = s - origin[i];
